backend/requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
pymupdf
python-docx
langgraph
python-dotenv
fastapi
uvicorn[standard]
python-multipart
requests
